fix(imagetool): harden controls bar teardown
Rely on QScrollArea's native sizing and keyboard focus navigation instead of installing the bar as an event filter on dynamic controls. This prevents PyQt6 teardown crashes while preserving responsive scrolling across dimensionality changes.
